Allow Content-Disposition settings like filename*=''...
At least one version of Microsoft Entourage produces Content-Disposition
headers like this one:

-- snip --
Content-disposition: attachment;
 filename*=''Study%20Informati%8Ar%20Singers_B.doc
-- snap --

Let's cope with them, too.

Unfortunately, it seems that there is no way to convince the DFA to handle

-- snip --
Content-Disposition: attachment;
 filename*=Imaris%207.4%20-%20Release%20Notes%20-%20draft.pdf
-- snap --

too (produced by ColdFusion 8 Application Server). The problem is that we
cannot have a rule with two branches one of which wants to copy to 'value'
and one of which wants to copy to 'minor'.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>